WIA/RVZ: Move (de)compression to a different file
