+++ Source/Core/DiscIO/DiscIO.vcxproj.filters
@@ -96,6 +96,9 @@
     <ClCompile Include="LaggedFibonacciGenerator.cpp">
       <Filter>Volume\Blob</Filter>
     </ClCompile>
+    <ClCompile Include="WIACompression.cpp">
+      <Filter>Volume\Blob</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="DiscScrubber.h">
@@ -174,6 +177,9 @@
       <Filter>Volume\Blob</Filter>
     </ClInclude>
     <ClInclude Include="LaggedFibonacciGenerator.h">
+      <Filter>Volume\Blob</Filter>
+    </ClInclude>
+    <ClInclude Include="WIACompression.h">
       <Filter>Volume\Blob</Filter>
     </ClInclude>
   </ItemGroup>
